chore: moved load_villain, changed koth training to validation in main.
