+++ poker/train.py
@@ -11,11 +11,8 @@ from collections import defaultdict
 import copy
 import time
 import logging
-
-from models.networks import OmahaActor,BetAgent
 from models.model_updates import update_actor_critic,update_combined,update_critic_batch,update_actor_critic_batch
 from utils.data_loaders import return_trajectoryloader
-from utils.utils import return_latest_baseline_path,return_next_baseline_path
 from models.model_utils import scale_rewards,soft_update,copy_weights,load_weights
 from tournament import tournament
 from db import MongoDB
@@ -26,15 +23,6 @@ def pad_state(state,maxlen):
     N = maxlen - state.shape[1]
     padding = np.zeros(N)
     return padded_state
-
-def load_villain(seed,nS,nA,nB,network_params,device,baseline_path):
-    baseline_path = return_latest_baseline_path(baseline_path)
-    if baseline_path:
-        villain = OmahaActor(seed,nS,nA,nB,network_params).to(device)
-        load_weights(villain,baseline_path)
-    else:
-        villain = BetAgent()
-    return villain
 
 def generate_vs_frozen(env,actor,critic,villain,training_params,id):
     # actor.eval()
@@ -257,10 +245,7 @@ def train_combined(env,model,training_params,learning_params,id):
         if e % training_params['save_every'] == 0 and id == 0:
             torch.save(model.state_dict(), os.path.join(training_params['save_dir'],f'OmahaCombined_{e}'))
 
-def train_dual(id,env,actor,critic,target_actor,target_critic,training_params,learning_params,network_params,validation_params):
-    # if validation_params['koth']:
-    #     villain = load_villain(seed,nS,nA,nB,network_params,learning_params['device'],training_params['baseline_path'])
-    villain = BetAgent()
+def train_dual(id,env,villain,actor,critic,target_actor,target_critic,training_params,learning_params,network_params,validation_params):
     for e in range(training_params['training_epochs']):
         sys.stdout.write('\r')
         if validation_params['koth']:
@@ -278,14 +263,3 @@ def train_dual(id,env,actor,critic,target_actor,target_critic,training_params,le
         if (e+1) % training_params['save_every'] == 0 and id == 0:
             torch.save(actor.state_dict(), os.path.join(training_params['actor_path'],f'OmahaActor_{e}'))
             torch.save(critic.state_dict(), os.path.join(training_params['critic_path'],f'OmahaCritic_{e}'))
-        #     # validate vs baseline
-        #     if validation_params['koth']:
-        #         results,stats = tournament(env,actor,villain,['hero','villain'],validation_params)
-        #         model_result = (results['trained_model']['SB'] + results['trained_model']['BB']) - (results['baseline_evaluation']['SB'] + results['baseline_evaluation']['BB'])
-        #         # if it beats it by 60%
-        #         if model_result  > (validation_params['epochs'] * .60):
-        #             # save weights as new baseline, otherwise keep training.
-        #             new_baseline_path = return_next_baseline_path(training_params['baseline_path'])
-        #             torch.save(actor.state_dict(), new_baseline_path)
-        #             # load new villain
-        #             villain = load_villain(seed,nS,nA,nB,network_params,learning_params['device'],training_params['baseline_path'])
